+++ src/aosm/HISTORY.rst
@@ -19,6 +19,7 @@ unreleased
 * Workaround Oras client bug (#90) on Windows for Artifact upload to ACR
 * Take Oras 0.1.18 so above Workaround could be removed
 * Take Oras 0.1.19 to fix NSD Artifact upload on Windows
+* Support deploying multiple instances of the same NF in an SNS
 
 0.2.0
 ++++++